+++ terrascript/resource/__init__.py
@@ -62,6 +62,7 @@ from .rundeck import *
 from .scaleway import *
 from .softlayer import *
 from .statuscake import *
+from .signalfx import *
 from .spotinst import *
 from .template import *
 from .tls import *