+++ terrascript/signalfx/__init__.py
@@ -1,0 +1,6 @@
+# terrascript/signalfx/__init__.py
+import terrascript
+
+
+class scaleway(signalfx.Provider):
+    pass