+++ terrascript/signalfx/d.py
@@ -1,0 +1,15 @@
+# terrascript/signalfx/d.py
+import terrascript
+
+
+class signalfx_aws_services(terrascript.Data):
+    pass
+
+class signalfx_azure_services(terrascript.Data):
+    pass
+
+class signalfx_dimension_values(terrascript.Data):
+    pass
+
+class signalfx_gcp_services(terrascript.Data):
+    pass